Move files, update namespaces
