+++ analyzers/src/SonarAnalyzer.Common/Analyzers/ParametrizedDiagnosticAnalyzer.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Analyzers
 {
     public abstract class ParametrizedDiagnosticAnalyzer : SonarDiagnosticAnalyzer
     {
