+++ analyzers/src/SonarAnalyzer.VisualBasic/Common/DescriptorFactory.cs
@@ -18,19 +18,12 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarAnalyzer.Helpers
+namespace SonarAnalyzer.Common
 {
-    public abstract class HotspotDiagnosticAnalyzer : SonarDiagnosticAnalyzer
+    internal static class DescriptorFactory
     {
-        protected IAnalyzerConfiguration Configuration { get; }
-
-        protected HotspotDiagnosticAnalyzer(IAnalyzerConfiguration configuration) =>
-            Configuration = configuration;
-
-        protected bool IsEnabled(AnalyzerOptions options)
-        {
-            Configuration.Initialize(options);
-            return SupportedDiagnostics.Any(d => Configuration.IsEnabled(d.Id));
-        }
+        public static DiagnosticDescriptor Create(string id, string messageFormat, bool? isEnabledByDefault = null, bool fadeOutCode = false) =>
+            // RuleCatalog class is created from SonarAnalyzer.SourceGenerator
+            DiagnosticDescriptorFactory.Create(AnalyzerLanguage.VisualBasic, RuleCatalog.Rules[id], messageFormat, isEnabledByDefault, fadeOutCode);
     }
 }
