+++ analyzers/src/SonarAnalyzer.VisualBasic/SonarAnalyzer.VisualBasic.csproj
@@ -34,6 +34,7 @@
     <Using Include="Microsoft.CodeAnalysis.VisualBasic.Extensions" />
     <Using Include="Microsoft.CodeAnalysis.Shared.Extensions" />
     <Using Include="SonarAnalyzer.AnalysisContext" />
+    <Using Include="SonarAnalyzer.Analyzers" />
     <Using Include="SonarAnalyzer.Common" />
     <Using Include="SonarAnalyzer.Extensions" />
     <Using Include="SonarAnalyzer.Helpers" />
