+++ analyzers/tests/SonarAnalyzer.Test/Analyzers/DiagnosticDescriptorFactoryTest.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarAnalyzer.Test.Helpers
+namespace SonarAnalyzer.Test.Analyzers
 {
     [TestClass]
     public class DiagnosticDescriptorFactoryTest
