+++ analyzers/tests/SonarAnalyzer.Test/SonarAnalyzer.Test.csproj
@@ -44,6 +44,7 @@
 	  <Using Include="Microsoft.CodeAnalysis" />
 	  <Using Include="Microsoft.CodeAnalysis.Diagnostics" />
     <Using Include="Microsoft.VisualStudio.TestTools.UnitTesting" />
+    <Using Include="SonarAnalyzer.Analyzers" />
     <Using Include="SonarAnalyzer.Common" />
     <Using Include="SonarAnalyzer.Helpers" />
     <Using Include="SonarAnalyzer.TestFramework.Analyzers" />
