+++ analyzers/tests/SonarAnalyzer.TestFramework/SonarAnalyzer.TestFramework.csproj
@@ -30,6 +30,7 @@
 	  <Using Include="Microsoft.CodeAnalysis" />
 	  <Using Include="Microsoft.CodeAnalysis.Diagnostics" />
     <Using Include="Microsoft.VisualStudio.TestTools.UnitTesting" />
+    <Using Include="SonarAnalyzer.Analyzers" />
     <Using Include="SonarAnalyzer.Common" />
     <Using Include="SonarAnalyzer.Helpers" />
     <Using Include="SonarAnalyzer.TestFramework.Common" />
